iqkms-signing: refactor into multiple modules
Factors each major type into its own module.

Also adds some additional lints.
