+++ iqkmsd/src/main.rs
@@ -3,7 +3,7 @@ use tonic::transport::Server;
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let addr = "[::1]:27100".parse().unwrap();
-    let keyring = signing::KeyRing::new();
+    let keyring = signing::Keyring::new();
     let eth_service = ethereum::SignerService::new(keyring);
 
     println!("Listening on {}", addr);
